airflow_docker/requirements.txt
snowflake-snowpark-python
python-dotenv
pandas
langchain
langchain-community
langgraph
pydantic
langchain-snowflake>=0.2.2
requests
zstandard
streamlit
snowflake-connector-python